Rename Binance2 -> Binance
